# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI4BMI API Groupe 3",
        "description": "API REST pour le système BMI — Gestion équipements et E-commerce BMI",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "description": "Vérifie les identifiants et retourne un Bearer Token. Valable pour tous les types de comptes.",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de connexion",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@bmi.bj"
                                    },
                                    "password": {
                                        "description": "Mot de passe sécurisé",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password_secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123def456"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "Jean Dupont"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@bmi.bj"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides (Email ou mot de passe incorrect)"
                    },
                    "422": {
                        "description": "Format des données invalide (ex: email non valide)"
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription client",
                "description": "Crée un nouveau profil utilisateur avec le rôle \"client\" par défaut.",
                "operationId": "6796c44e1f6079c411abbfed066c5907",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "fullname": {
                                        "description": "Nom complet du client",
                                        "type": "string",
                                        "example": "Client Test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "client@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "97000000",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Cotonou, Bénin",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte client créé avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation ou email déjà utilisé"
                    }
                }
            }
        },
        "/api/v1/admin/users": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Créer un compte personnel (Admin only)",
                "description": "Permet à un administrateur authentifié de créer un technicien ou un gestionnaire.",
                "operationId": "6ab0e37ebd7fec25d941864469045645",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Jean Technicien"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean@bmi.bj"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "97000000"
                                    },
                                    "role": {
                                        "description": "Le rôle professionnel à attribuer",
                                        "type": "string",
                                        "enum": [
                                            "technicien",
                                            "gestionnaire"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte personnel créé avec succès"
                    },
                    "401": {
                        "description": "Session non authentifiée"
                    },
                    "403": {
                        "description": "Accès interdit - Droits administrateur requis"
                    },
                    "422": {
                        "description": "Données invalides ou email déjà pris"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion",
                "description": "Invalide le jeton d'accès actuel pour déconnecter l'utilisateur proprement.",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Session fermée avec succès"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Profil utilisateur connecté",
                "description": "Récupère les détails du compte de la session active (Id, Nom, Email, Téléphone, Adresse, Rôle).",
                "operationId": "ae2f8c1ae1f9c5aee99c01f5d13da30b",
                "responses": {
                    "200": {
                        "description": "Profil récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "Jean Dupont"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@bmi.bj"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentification requise"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
}